bertopic/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
args==0.1.0
bertopic==0.16.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
clint==0.5.1
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.1.1
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
de-core-news-sm==3.7.0
distro==1.9.0
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
germansentiment==1.1.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hdbscan==0.8.33
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.1
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==7.0.1
importlib-resources==6.1.2
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
llvmlite==0.41.1
lxml==5.1.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
nltk==3.8.1
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai==1.13.3
packaging==23.2
pandas==2.0.3
pillow==10.2.0
plotly==5.19.0
pluggy==1.4.0
preshed==3.0.9
pydantic==2.6.1
pydantic-core==2.16.2
pynndescent==0.5.11
pyparsing==3.1.1
pytest==8.1.1
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rfc3986==1.5.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
textplot==0.1.1
thinc==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
tomli==2.0.1
torch==2.2.0
tqdm==4.66.2
transformers==4.37.2
triton==2.2.0
typer==0.9.0
typing-extensions==4.9.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==2.2.0
wasabi==1.1.2
weasel==0.3.4
zipp==3.17.0
